+++ photo_hdr_duval_BD/2_S2_createAgent.sql
@@ -7,8 +7,8 @@ CREATE TABLE Agent.Agent(
 	NomPrenomAgent NVARCHAR(50) NOT NULL,
 	NomEntreprise NVARCHAR(50) NOT NULL,
 	Adresse NVARCHAR(50) NOT NULL,
-	TelPrincipal INT NOT NULL,
-	TelSecondaire INT NOT NULL
+	TelPrincipal BIGINT NOT NULL,
+	TelSecondaire BIGINT NOT NULL
 
 	PRIMARY KEY (AgentID)
 ) ON [PRIMARY];
